move exercises functions, add code for section 4 and solve the new exercises
