+++ cmd/go_learning.go
@@ -44,6 +44,9 @@ func main() {
 		"chapter01.Dup2":       chapter01.Dup2,
 		"chapter01.Dup3":       chapter01.Dup3,
 		"chapter01.Exercise04": chapter01.Exercise04,
+		"chapter01.Lissajous":  chapter01.Lissajous,
+		"chapter01.Exercise05": chapter01.Exercise05,
+		"chapter01.Exercise06": chapter01.Exercise06,
 	}
 
 	f := functions[*functionNamePtr]
