+++ pkg/chapter01/04_lissajous.go
@@ -1,0 +1,94 @@
+// Package chapter01 04_lissajous contains the code studied on section 04
+package chapter01
+
+import (
+	"image"
+	"image/color"
+	"image/gif"
+	"io"
+	"math"
+	"math/rand"
+	"os"
+	"time"
+)
+
+// Lissajous creates GIF animations for lissajous random figures
+func Lissajous() {
+	palette := []color.Color{color.White, color.Black}
+	rand.Seed(time.Now().UTC().UnixNano())
+	lissajous(os.Stdout, palette)
+}
+
+// lissajous implements the GIF creator
+func lissajous(out io.Writer, palette []color.Color) {
+	const (
+		cycles  = 5
+		res     = 0.001
+		size    = 100
+		nframes = 64
+		delay   = 8
+	)
+
+	freq := rand.Float64() * 3.0
+	anim := gif.GIF{LoopCount: nframes}
+	phase := 0.0
+	for i := 0; i < nframes; i++ {
+		rect := image.Rect(0, 0, 2*size+1, 2*size+1)
+		img := image.NewPaletted(rect, palette)
+		for t := 0.0; t < cycles*2*math.Pi; t += res {
+			x := math.Sin(t)
+			y := math.Sin(t*freq + phase)
+			img.SetColorIndex(size+int(x*size+0.5), size+int(y*size+0.5), 1)
+		}
+		phase += 0.1
+		anim.Delay = append(anim.Delay, delay)
+		anim.Image = append(anim.Image, img)
+	}
+	gif.EncodeAll(out, &anim)
+}
+
+// Exercise05 changes the lissajous fuction color palette
+func Exercise05() {
+	palette := []color.Color{color.Black, color.RGBA{0x00, 0xff, 0x00, 0xff}}
+	rand.Seed(time.Now().UTC().UnixNano())
+	lissajous(os.Stdout, palette)
+}
+
+// Exercise06 changes the lissajous fuction to alternate the color index
+func Exercise06() {
+	palette := []color.Color{
+		color.Black,
+		color.RGBA{0xff, 0x00, 0x00, 0xff},
+		color.RGBA{0x00, 0xff, 0x00, 0xff},
+		color.RGBA{0x00, 0x00, 0xff, 0xff},
+	}
+	rand.Seed(time.Now().UTC().UnixNano())
+	lissajousAlternative := func(out io.Writer, palette []color.Color) {
+		const (
+			cycles  = 5
+			res     = 0.001
+			size    = 100
+			nframes = 64
+			delay   = 8
+		)
+
+		freq := rand.Float64() * 3.0
+		anim := gif.GIF{LoopCount: nframes}
+		phase := 0.0
+		for i := 0; i < nframes; i++ {
+			rect := image.Rect(0, 0, 2*size+1, 2*size+1)
+			img := image.NewPaletted(rect, palette)
+			for t := 0.0; t < cycles*2*math.Pi; t += res {
+				x := math.Sin(t)
+				y := math.Sin(t*freq + phase)
+				img.SetColorIndex(size+int(x*size+0.5), size+int(y*size+0.5),
+					uint8(rand.Intn(len(palette)-1)+1))
+			}
+			phase += 0.1
+			anim.Delay = append(anim.Delay, delay)
+			anim.Image = append(anim.Image, img)
+		}
+		gif.EncodeAll(out, &anim)
+	}
+	lissajousAlternative(os.Stdout, palette)
+}
